fix: rendering, approval bug
